Added dragons to to the project and got the main camera to follow dragon+
Also added Signal Emitters that disable the dollies, Far and Near Spline.
I have 3 cameras a far spline camera, a near spline camera and an overview camera; with blends between them.
I have added an empty object which I have named Director. I have added a Playable Director to this and a Timeline (which you have to lock to ensure that it is this timeline you are working with).
Set up the animations for the dragon, with the flying clip and initial clip of him above the house the run clip, then wait clip before skill 2 clip (which is him shooting) before returning to wait.